Allow substituting from different storeDir
Substituters can substitute from one store dir to another with a
little bit of help. The store api just needs to have a CA so it can
recompute the store path based on the new store dir. We can only do
this for fixed output derivations with no references, though.


Former-commit-id: 79c169d1c615211af69c1cbc6218f7465a4f81ed