Fix SRI test
We can’t use custom name here because different names will have
different store paths. This is a limitation of the Store API’s
reliance on store paths.

We might be able to get around the above in the future by using a
dummy name for certain fixed output paths.


Former-commit-id: 006f1252d2fdb36c7ee66bf79f0df2ab8cd16816