Make `Buildable` a `std::variant`
I think this better captures the intent of what's going on: we either
have an opaque store path, or a drv path with some outputs.

Having this structure will also help us support CA derivations: we'll
have to allow the outpath paths to be optional, so the structure we gain
now makes up for the structure we loose then.


Former-commit-id: e1de1fe0d82d8ba702947dcad3b678cbb9ce9333