Fix max-jobs option
After e487a974de4d095c406c1be6820c02ed08e2b42d [formerly 0ed946aa616bbf7ffe7f90d3309abdd27d875b10], max-jobs setting (-j/--max-jobs)
stopped working.

The reason was that nrLocalBuilds (which compared to maxBuildJobs to figure
out whether the limit is reached or not) is not incremented yet when tryBuild
is started; So, the solution is to move the check to tryLocalBuild.

Closes https://github.com/nixos/nix/issues/3763


Former-commit-id: ed218e1d6cf755fc3011c0954eb7031f95d3d732