configure: explicit dependency on lowdown library
This dependency is used from quite a long time (now in libcmd) but
was not explicitly stated in the configure phase, possibly leading
to quite late build failures if that was not met (ie. building it
outside the .nix files provided). This MR adds it in the configure
phase so the failure is early and error is much more explicit.


Former-commit-id: e9dbba0fc7ab90e3679ba57245b36420ce76ccc4