Don't reset the logger in a vfork
41ffe85314c7bdd46724b1dae56694a19d144b11 [formerly 9c766a40cbbd3a350a9582d0fd8201e3361a63b2] broke logging from the
daemon, because commonChildInit is called when starting the build hook
in a vfork, so it ends up resetting the parent's logger. So don't
vfork.

It might be best to get rid of vfork altogether, but that may cause
problems, e.g. when we call an external program like git from the
evaluator.


Former-commit-id: c6718a9d95021476f0cca47b61479e4046090f4e