Make init.sh safe to run twice
`init.sh` is tested on its own. We used to do that. I deleted it in
0bd3fd4d8d82fbf502e13bbd098b952a6bd3a570 [formerly 4720853129b6866775edd9f90ad6f10701f98a3c] but I am not sure why. Better
to just restore it; at one point working on this every other test
passed, so seems good to check whether `init.sh` can be run twice.

We don't *need* to run `init.sh` twice, but I want to try to make our
tests as robust as possible so that manual debugging (where tests for
better or worse might be run ways that we didn't expect) is less
fragile.


Former-commit-id: 5dbbf233327bff1c89088ed0fc45882563240b1a