Start the pager early-enough in `nix why-depends`
`nix why-depends` is piping its output into a pager by default.
However the pager was only started after the first path is printed,
causing it to be excluded from the pager output.

(Actually the pager was started *inside* the recursive function that was
printing the dependency chain, so a new instance was started at each
level. It’s a little miracle that it worked at all).

Fix #5911


Former-commit-id: 81cd0a113b8da1485c2f892c9dd9edb541f139ad